extract labs and tests methods into adapter, adjust serializer + usage, update all tests
