+++ lib/unified_health_data/models/lab_or_test.rb
@@ -22,9 +22,10 @@ module UnifiedHealthData
     attribute :sample_tested, String
   end
 
-  class Attributes
+  class LabOrTest
     include Vets::Model
 
+    attribute :id, String
     attribute :display, String
     attribute :test_code, String
     attribute :date_completed, String
@@ -34,13 +35,5 @@ module UnifiedHealthData
     attribute :ordered_by, String
     attribute :body_site, String
     attribute :observations, UnifiedHealthData::Observation, array: true
-  end
-
-  class LabOrTest
-    include Vets::Model
-
-    attribute :id, String
-    attribute :type, String
-    attribute :attributes, UnifiedHealthData::Attributes, array: false
   end
 end
